refactor: add `roco` namespace to `mw`
